Detect duplicate completions on TCP/IP stream.
For some reason Windows sometimes sends a send page completion
twice for the same object. This leads to bug checks because
of double free's and other things. We now detect those double
frees by keeping the completions on a list and refusing to
free the resources again if the completion is not on the
list any more. A bug counter tells us when the bug occurs
(approximately once a day).