Make nr_requests comparision signed.
Else it may fail (and hang forever in inc_ap_bio) when nr_request
becomes negative. nr_requests being negative is an unrelated bug
(which happens for example when a network connection is interrupted)
and should be fixed one day.

This now solves the problem that a Windows booted via WinDRBD survives
and can recover a failed connection (without performance impact
when connected again).