Drbd.cgi: retry forever on error.
When rebooting a Windows machine, it stays primary connected
for a while (depending on Ping timeout). That causes accessing
the DRBD device from the CGI script to fail (temporarily) with
Wrong media type.

With this patch, the drbd.cgi script repeats forever (with a
sleep 1 in it) to try to deliver the data (and get the block
size).