Fixed IRQL bug.
flags2 is PASSIVE_LEVEL, flag is DISPATCH_LEVEL. When restoring
in wrong order we leave queue_work with IRQL=DISPATCH_LEVEL which
is not good.

Now, disconnect while sync test ran for 250 iterations (then BSODed,
also use after free bugs in there). We need to disallow new work
to be queued when we are already in destroy_workqueue().

(converted sources hash is 26d8917edb698d