DRBD patches for speeding up resync when sync target.
We use 'big pages': since pages are regular memory allocated by
kmalloc() we can have pages of any size. With this patch we do not
split resync requests > 4K into 4K blocks (which would be submitted
one per one to the backing device, which makes it slow). Instead
we handle the resync request with one big pages (and also submit
that one big page to the backing device).