Several attempts to fix receiver cache.
One is that sometimes receiver cache is empty and kernel_recvmsg
gets woken up. In that case we ignore the request (set to 0)
TODO: there should be a continue ...

Also prints err returned by wsk_recvmsg() .. we shouldn't exit the
receiver thread if this is just a temporary (EAGAIN, ...) error.