Fix: wait_event[_timeout] should not be interruptible.
In the disconnect while sync test there was a BSOD because
wait_event in actlog (wait_until_done_or_force_detached())
returned early with a not handled error code (-EINTR).

We now go and check if this patch fixes the BSOD (and it
should also fix the disk going to Failure bug).