Recursion protection for rcu_read_lock
This might fix the freeze. What we had was:

Thread A	Thread B
rcu_read_lock()
		synchronize_rcu()
rcu_read_lock()

Where synchronize_rcu wait for rcu_read_lock #1 and rcu_read_lock #2
waits forever for synchronize_rcu.

We do not take the RCU lock again (not even shared) when it is
already taken in the thread.