put_page in bio_finished if WRITE ...
Not sure about this one ... we still get BSODs from upper writes
but this time in the free routine of the caller ... somebody does
not clean up everything in time. Probably we want to complete writes
in free_mdls thread once the MDLs are freed (something like
IoCompleteRequest(bio->upper_irp) ...) but how did this work
before?