bio_clone: take a reference to original bio.
DRBD uses this solely to create a lower disk I/O bio from the
upper bio (as passed in drbd_make_request()). The reference
is dropped when the cloned bio (lower bio) is freed but after
the lower bio is inserted in the to_be_freed list. This way,
the upper bio is always freed *after* all lower bios are
freed.

We think that this should solve the PTE BSOD we observe.