Disable debug output.
With setting the UniqueID field to TRUE we make the Device ID's
unique on the system without PnP manager adding stuff to the
device instance path.

Starting with this commit the format of the device instance path
has changed so if you depend on the device instance path you should
note that it has changed now.