Fixed cast in IS_ERR_VALUE macro.
Starting with a recent commit the ERR_PTR (and related) functions
use LONG_PTR (which is 64 bit on 64 bit Windows and 32 bit on 32
bit Windows, long is always 32 bit!). We now also corrected this
in the IS_ERR_VALUE macro. This patch should fix the BSOD on specifying
an unsupported (i.e. != crc32c) crypto hash on drbdadm down (code
path did not recognize the error)