Reworked remove_lock in windrbd_device layer.
There was a (most likely) bug: we used the remove_lock from
the bdev instead of the w_remove_lock of the device extension
to wait for I/O to complete. This lead to NULL pointer dereferences
in SCSI and windrbd_finish_bio() and should be fixed now.
Also we are not unsetting the ref pointer (but I think we
could). Also we use the device extension (ref) directly
in windrbd_scsi and windrbd_io.