Stop log service after driver
So we see the last log messages. This was a bit complicated
we had to remove the service dependencies from the registry
and if they were there stop logger before the driver else
after the driver.

Now driver is not loaded at boot time, this has to be done
manually until we also patch the start type from 3 to 1
in the registry.