Debug printk's for researching 16TB bug.
It looks like changing sector_count from some 32 bit value
to 64 bits breaks the REVERSE_BYTES() assignment (upper
4 bytes remain uninitialized). Solution is probably
initializing sector_count first (to 0) and then do the
REVERSE_BYTES(). Unfortunately this macro doesn't even
check for target endianess (such as big endian as we
have it on ARM) but this is a Microsoft problem.