inno-setup: ExecWithResult: set an optional PATH component before running a program.
This should enable running the installer from a Windows
background service. The problem was that programs that
are not in the {app} (usually C:\Program Files\WinDRBD)
path (like windrbd.exe, drbdadm.exe, ...) failed to run
when the installer is started from a service. The service
does not get the Environment from the HKLM\..\Session Manager
key, so adding PATH components there does nothing.