Yes ResilienceDeviceProperty was invalid.
For now we don't support reporting it (it is only partially
documented). Reason is that we would have to track how
many physical (UpToDate) copies there are which might
change from time to time. Reporting a zero length struct
breaks ReFS compatibility, so we just pretend we do not
support the query.