drbd_main and master plan how to squeeze patches.
We are much faster when we start all over again and
diff the current converted sources with the original
DRBD sources we base on. The alternative would be
to review the about 300 patch files if we still need
them which would take much longer (and is less precise
and less fun to do).

Also including the file workflow which explains the
meta comment categories.