Re-generate virtual partition table on device size change.
This does NOT reread the boot sector from DRBD since
during device size change DRBD suspends I/O (and therefore
we run into a timeout and later a BSOD). So the boot
sector will be cached.

Apart from that I would suggest a spinlock around the
in-memory generation / returning via SCSI requests
just to be sure we don't crash there.

This fixes an issue with resizing an NTFS partition
on a DRBD device with virtual partition tables (after
secondary / primary the DRBD device came up as RAW
(instead of NTFS).