Do not change size of partition.
Windows (and most likely Linux, too) doesn't like that
the size of a partition is changed by the storage driver
without notice. So keep the size of the partition on
drbdadm resize and have diskpart do the rest.

On rescanning Windows dismounted and mounted again the
volume which caused running servives (databases, ...) to
crash and/or behave in an unwanted way. This commit should
fix this behavior. The procedure now would be:

(diskpart) select volume W # backing device
(diskpart) extend size=1024 # in MB
(shell) drbdadm resize res
(diskpart) rescan
(diskpart) select volume D # the partition on the WinDRBD disk
(diskpart) extend # to maximum size