I/O cork/uncork: fix submission of standalone bio.
We check if sectors to write are adjacent by looking at
the next bio. If not and if the last bio was the only
bio within his join group it is submitted directly
without creating a joined bio. This commit fixes
a bug where the followup bio instead of the last bio
was submitted to the lower device. This also fixes
deletion of the correct bio from the list (which
should fix random memory corruption we observed).