WinDRBD API function to complete all in-flight bios.
When a disk timeout in DRBD happens we want to be able
to call bi_endio() on all bios that are currently waiting
for the disk with a failure status (something like timeout).

However it might happen that the disk responds after the
timeout elapsed. So we need to make sure that bi_endio()
(in DRBD) is only called once.

This patch adds support for that in WinDRBD. The function
should be called in the disk timeout handler (request_timer_fn).