Makefile: cocci for new transformation method.
Currently 2 cocci scripts are in place. A directory
drbd-tmp is created (do not start coding in there
it is removed by make clean).
The old one (converted-sources) will go away 'soon'
(as soon as the new one compiles and links)
Also (tries to) compile drbd-tmp but currently there
are about 2500 compilation errors.