Release 1.2.0-rc4
Now the old patching mechanism via patch files has gone. It is
replaced by some (minimal) patches on DRBD itself (on a separate
branch) plus some generic cocci scripts. The majority of the
work was to be as Linux compatible (headers, ...) as possible
in order to minimize patches to DRBD.

What is patched: mainly the RCU / spinlocks / some initialization
(this might go away soon). Everything else is in the compat layer
('WinDRBD').