Define WinDRBDNonPagedPool
Instead of using the Microsoft mechanism (one binary for
'old' and 'new') (which does not exist in our headers),
we have the WinDRBDNonPagedPool define that is
NonPagedPoolNx for 'newer' Windows and 'NonPagedPool'
for 'old' Windows (if REACTOS is defined).

That reduced the NonPagedPool allocations from 700 something
to 7. I assume those 7 are from the NPagedPool lookaside
lists. I will replace them by ordinary kmalloc's in the
next commit.