Relaxing CPU at various places...
TCP/IP stack starvation. We get 5000+ send requests in the
queue which are never sent. Eventually a PingAck didn't arrive
in time happens and the close_socket() will cause the send
requests being completed with an error.

Maybe this is *because* of the logging?

msleep'ing in wsk_sendpage did the trick..