DRBD 9.2 almost compiles and links.
Commented out TLS in the DRBD 9.2, we will probably never
support this. If you need encryption on the wire use an
ssh or TLS tunnel.

Also use the work variant for handling data received callbacks.
The other implementation (using tcp_read_sock and several
skb_xxx() functions) was too deep inside the Linux kernel
to be implemented in WinDRBD.