Fixed cancel_work_sync() implementation
Well ... sort of.

We wait for all work on the queue to finish and do not call
the work that is cancelled. It does the same but a bit (ok
a lot) more than what was asked for (cancel all work instead
of cancelling one work).

drbdadm disconnect <res> now does not BSOD.

One day we need to reimplement (or take Linux implementation)
the whole work queue mess ...