Some minor fixes.
may_printk is 0 if it is a UDP socket. Maybe we should get
rid of this flag completely now that we have recursion detection
in printk().

Also sock_sendmsg calls do_send() directly now. I think it should
also pass the page as parameter so the page gets retained (get_page/
put_page).