Makefile: honor DEFAULT target on checking for building dependencies
The problem is that if Makefile includes the dependecies they
are built outside the docker image (if default is package-in-docker)
with whatever cocci is installed there. This patch uses the
(existing) DEFAULT variable in case no target is given at the
command line. So if one just types make the docker build is
started without creating the dependencies on the host system
first.