Do not handle BusRelations for disk devices.
On ReactOS (and probably also on Windows Server 2003) the partition
manager fills in partitions as child device objects of the disk.
If we overwrite this and pretend we have no children partition
manager (and pnp manager) fail to signal a pnp event so that
no volume objects gets created. Thereby no drive letter can
be assigned.

This patch should fix that.