SCSI_INQUIRY has different pages.
Page 0 tells the host which pages are supported.
Page 0x83 is the actual device identification struct.
Don't know if we need other pages as well but probably
yes.

sg_vpd is a useful tool under Linux. Also on Windows
PowerShell prints the vendor/model strings with Get-Disk | Format-All
or so ...