Copy page contents to tmp buffer before sending it.
At least on Windows Server 2003 SP2 using the page directly
runs into a BSOD quite soon. We might want to add conditional
compilation (like #ifdef WINNT_52) since it is not an issue
on modern windows (and might slow down I/O because of double
buffering).

It looks as if this fixes the PFN list corrupt BSOD.