Researching reason for pfn list corrupt.
It looks as if Windows Server 2003 SP2 kernel gets confused
when the locked region is not the same as the allocated region
(we have KMALLOC_DEBUG defined which adds 256 (242 on 32 bit)
bytes *before* the pointer returned (and the pointer returned
is used to feed the Mdl)).

We now try without kmalloc_debug().