User mode helper: check if daemon is still running while waiting for the return value.
If the user mode helper daemon fetches a user mode helper request
and terminates before it sends the return value of the user mode
helper command the kernel will wait forever on the return value.
The fix is to also check if the user mode helper daemon is running
while waiting for the return value.