Sleep 1 second before closing a socket.
In our tests this has fixed the BrokenPipe transistion. It
mitigates a bug in our implementation (or the Windows networking
stack) that cases the last packet not to be delivered.
This also happens when testing without DRBD being involved
(see the send_a_lot() test in windrbd_test.c, 1 out of 1000
files were missing the last few bytes).

With this commit, all tests flagged for WinDRBD readyness
(about 25% of the total number of tests) succeed.